fix bug in process_button(), refactor Lights_SetState()
open to suggestions on formatting/style
for functions with several parameters, whether to follow the style of CALLBACK_HID_Device_CreateHIDReport() or the way I did it when calling process_button() and set_led()
also I decided on snake_case for process_button(), set_led(), and fields in the button_pins. It just felt weird to use camelCase for setLed() or field led_pin since LED is an abbreviation. Would like your feedback on this